git-svn-id: https://www.seasar.org/svn/s2container.net/trunk/s2container.net@236 6374060a-8204-0410-be21-af3d0958e6ef

+++ source/Seasar.Tests/Seasar.Tests.Extension.DataSets/Impl/XlsReaderTest1.cs
@@ -48,9 +48,9 @@ namespace Seasar.Tests.Extension.DataSets.Impl
 				Assert.AreEqual(4, dataSet.Tables["Type"].Rows.Count);
 
 				Assert.AreEqual(typeof(string), dataSet.Tables["Type"].Columns["stringValue"].DataType);
-				Assert.AreEqual(typeof(decimal), dataSet.Tables["Type"].Columns["intValue"].DataType);
-				Assert.AreEqual(typeof(decimal), dataSet.Tables["Type"].Columns["longValue"].DataType);
-				Assert.AreEqual(typeof(decimal), dataSet.Tables["Type"].Columns["decimalValue"].DataType);
+				Assert.AreEqual(typeof(double), dataSet.Tables["Type"].Columns["intValue"].DataType);
+				Assert.AreEqual(typeof(double), dataSet.Tables["Type"].Columns["longValue"].DataType);
+				Assert.AreEqual(typeof(double), dataSet.Tables["Type"].Columns["decimalValue"].DataType);
 				Assert.AreEqual(typeof(double), dataSet.Tables["Type"].Columns["floatValue"].DataType);
 				Assert.AreEqual(typeof(DateTime), dataSet.Tables["Type"].Columns["dateTimeValue"].DataType);
